% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate-nc-data.R
\name{extract_nc_values}
\alias{extract_nc_values}
\title{Read a nc grid at specific sites}
\usage{
extract_nc_values(x, sites, x_var, y_var, t_var, v_var,
  convert_0_to_360 = FALSE, t_unit = "days", t_origin)
}
\description{
Read a nc grid at specific sites
}

